Break up stdlib
